Develop : birth 추가 + 파트너 조회 오류 수정 + 쿠키 오류 수정 + 프로필 사진 오류 수정 등

+++ run/src/main/java/com/guide/run/event/service/EventMatchingService.java
@@ -1,5 +1,7 @@
 package com.guide.run.event.service;
 
+import com.guide.run.attendance.entity.Attendance;
+import com.guide.run.attendance.repository.AttendanceRepository;
 import com.guide.run.event.entity.EventForm;
 import com.guide.run.event.entity.dto.response.form.Form;
 import com.guide.run.event.entity.dto.response.match.*;
@@ -11,6 +13,7 @@ import com.guide.run.partner.entity.matching.Matching;
 import com.guide.run.partner.entity.matching.UnMatching;
 import com.guide.run.partner.entity.matching.repository.MatchingRepository;
 import com.guide.run.partner.entity.matching.repository.UnMatchingRepository;
+import com.guide.run.partner.service.PartnerService;
 import com.guide.run.user.entity.type.UserType;
 import com.guide.run.user.entity.user.User;
 import com.guide.run.user.repository.user.UserRepository;
@@ -29,6 +32,9 @@ public class EventMatchingService {
     private final UserRepository userRepository;
     private final EventRepository eventRepository;
     private final EventFormRepository eventFormRepository;
+    private final PartnerService partnerService;
+    private final AttendanceRepository attendanceRepository;
+
     @Transactional
     public void matchUser(Long eventId, String viId, String userId) {
         eventRepository.findById(eventId).orElseThrow(NotExistEventException::new);
@@ -38,6 +44,9 @@ public class EventMatchingService {
         EventForm guideForm = eventFormRepository.findByEventIdAndPrivateId(eventId, guide.getPrivateId());
         Matching existGuide = matchingRepository.findByEventIdAndGuideId(eventId, guide.getPrivateId());
         if(existGuide!=null){
+            //existGuide의 파트너 삭제 추가
+            partnerService.setNotAttendGuidePartner(eventId, guide);
+
             matchingRepository.delete(existGuide);
             long countMatchedGuideForVi = matchingRepository.countByEventIdAndViId(eventId, existGuide.getViId());
             if(countMatchedGuideForVi==0){
@@ -48,8 +57,16 @@ public class EventMatchingService {
                                 .privateId(unMatchingVi.getPrivateId())
                                 .build()
                 );
+
+                //unMatching vi의 파트너 삭제 추가
+                partnerService.setNotAttendViPartnerList(eventId, unMatchingVi);
+
             }
+
+
         }
+
+
         matchingRepository.save(
                 Matching.builder()
                         .eventId(eventId)
@@ -66,9 +83,14 @@ public class EventMatchingService {
                         .build()
         );
         Optional<UnMatching> findVi = unMatchingRepository.findByPrivateIdAndEventId(vi.getPrivateId(),eventId);
+
         if(!findVi.isEmpty()){
             unMatchingRepository.delete(findVi.get());
         }
+
+        //출석 되어 있으면 파트너 반영.
+        updatePartnerOnAttendance(eventId, guide);
+
 
     }
 
@@ -77,8 +99,12 @@ public class EventMatchingService {
         eventRepository.findById(eventId).orElseThrow(NotExistEventException::new);
         User user = userRepository.findUserByUserId(userId).orElseThrow(NotExistUserException::new);
         String privateId = user.getPrivateId();
+
         if(user.getType()==UserType.VI){
             List<Matching> allMatching = matchingRepository.findAllByEventIdAndViId(eventId, privateId);
+            //vi 파트너 삭제 추가
+            partnerService.setNotAttendViPartnerList(eventId, user);
+
             for(Matching m : allMatching){
                 matchingRepository.delete(m);
                 unMatchingRepository.save(
@@ -95,8 +121,15 @@ public class EventMatchingService {
                             .eventId(eventId)
                             .build()
             );
+
+
+
         }else{
             Matching m = matchingRepository.findByEventIdAndGuideId(eventId, privateId);
+
+            //파트너 삭제 추가
+            partnerService.setNotAttendGuidePartner(eventId, user);
+
             matchingRepository.delete(m);
             unMatchingRepository.save(
                     UnMatching.builder()
@@ -206,30 +239,46 @@ public class EventMatchingService {
             Form guideForm = guideList.get(0);
             User guide = userRepository.findUserByUserId(guideForm.getUserId()).orElseThrow(NotExistUserException::new);
             User vi = userRepository.findUserByUserId(viList.get(i).getUserId()).orElseThrow(NotExistUserException::new);
-            matchingRepository.save(
-                    Matching.builder()
-                            .eventId(eventId)
-                            .guideId(guide.getPrivateId())
-                            .viId(vi.getPrivateId())
-                            .viRecord(viList.get(i).getApplyRecord())
-                            .guideRecord(guideForm.getApplyRecord())
-                            .build()
-            );
-            unMatchingRepository.delete(
-                    UnMatching.builder()
-                            .eventId(eventId)
-                            .privateId(guide.getPrivateId())
-                            .build()
-            );
-            Optional<UnMatching> findVi = unMatchingRepository.findByPrivateIdAndEventId(vi.getPrivateId(),eventId);
-            if(!findVi.isEmpty()){
-                unMatchingRepository.delete(findVi.get());
+
+            Matching existMatching = matchingRepository.findByEventIdAndGuideId(eventId, guide.getPrivateId());
+            if(existMatching==null) {
+                matchingRepository.save(
+                        Matching.builder()
+                                .eventId(eventId)
+                                .guideId(guide.getPrivateId())
+                                .viId(vi.getPrivateId())
+                                .viRecord(viList.get(i).getApplyRecord())
+                                .guideRecord(guideForm.getApplyRecord())
+                                .build()
+                );
+                unMatchingRepository.delete(
+                        UnMatching.builder()
+                                .eventId(eventId)
+                                .privateId(guide.getPrivateId())
+                                .build()
+                );
+
+                Optional<UnMatching> findVi = unMatchingRepository.findByPrivateIdAndEventId(vi.getPrivateId(),eventId);
+                if(!findVi.isEmpty()){
+                    unMatchingRepository.delete(findVi.get());
+                }
             }
+
+            //출석 되어 있으면 파트너 반영.
+            updatePartnerOnAttendance(eventId, guide);
+
             guideList.remove(0);
             startIdx++;
         }
         return startIdx;
     }
+
+    private void updatePartnerOnAttendance(Long eventId, User guide) {
+        Attendance attendance = attendanceRepository.findByEventIdAndPrivateId(eventId, guide.getPrivateId());
+        if(attendance != null && attendance.isAttend()) {
+            partnerService.setAttendGuidePartner(eventId, guide);
+        }
+        }
 
     /*
     public int autoMatchBetweenTwoGroup(List<Form> vi,List<Form> guide){
@@ -237,4 +286,6 @@ public class EventMatchingService {
         return vi.size();
     }
     */
+
+
 }
